Move unit test to its own file
