+++ src/test/java/com/powsybl/openloadflow/ac/AcLoadFlowTwoBusNetworkTest.java
@@ -16,14 +16,14 @@ import com.powsybl.loadflow.LoadFlowResult;
 import com.powsybl.math.matrix.DenseMatrixFactory;
 import com.powsybl.openloadflow.OpenLoadFlowParameters;
 import com.powsybl.openloadflow.OpenLoadFlowProvider;
-import com.powsybl.openloadflow.ac.nr.StateVectorScalingMode;
 import com.powsybl.openloadflow.network.SlackBusSelectionMode;
 import com.powsybl.openloadflow.network.TwoBusNetworkFactory;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
 import static com.powsybl.openloadflow.util.LoadFlowAssert.*;
-import static org.junit.jupiter.api.Assertions.*;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertTrue;
 
 /**
  * @author Geoffroy Jamgotchian <geoffroy.jamgotchian at rte-france.com>
@@ -66,28 +66,6 @@ class AcLoadFlowTwoBusNetworkTest {
         assertReactivePowerEquals(1.683, line1.getTerminal1());
         assertActivePowerEquals(-2, line1.getTerminal2());
         assertReactivePowerEquals(-1, line1.getTerminal2());
-    }
-
-    @Test
-    void withHighLoadTest() {
-        // TODO a better ill-conditioned case for state vector scaling tests, also to be moved to another test file
-        network.getLoad("l1").setP0(3.902); // 3.9 does not need scaling
-
-        LoadFlowResult result = loadFlowRunner.run(network, parameters);
-        assertFalse(result.isOk());
-        assertEquals(LoadFlowResult.ComponentResult.Status.MAX_ITERATION_REACHED, result.getComponentResults().get(0).getStatus());
-
-        parametersExt.setStateVectorScalingMode(StateVectorScalingMode.MAX_VOLTAGE_CHANGE);
-        result = loadFlowRunner.run(network, parameters);
-        assertTrue(result.isOk());
-        assertEquals(20, result.getComponentResults().get(0).getIterationCount());
-        assertVoltageEquals(0.6364204826103471, bus2);
-
-        parametersExt.setStateVectorScalingMode(StateVectorScalingMode.LINE_SEARCH);
-        result = loadFlowRunner.run(network, parameters);
-        assertTrue(result.isOk());
-        assertEquals(6, result.getComponentResults().get(0).getIterationCount());
-        assertVoltageEquals(0.6364204826103471, bus2);
     }
 
     @Test
